Add files via upload
This is random quote machine based on HTML,CSS,Bootstrap and JavaScript. In this mini project we can get the quote randomly.When you click on the button you will be getting the random quote with randomly background colors.So , I used here mainly html css and bootstrap for responsive design and one of the most popular Clint-side language JavaScript.